042stab111.12
    042stab111.12
    109.12 since 042stab109.10:
    * l2tp: Fix UDP socket reference count bugs in the pppol2tp (#3233)
    * ploop: added into pvd header new close signatures (PSBM-33882)
    * vmscan,tmpfs: treat used once pages on tmpfs as used once
      (PSBM-31450, PSBM-31450)
    * blk: add changed block trace infrastructure (PSBM-32023)
    * block: cbt fixup use-after-free inside __blk_cbt_set
    * cbt: Use propper mem allocation context
    * cbt: support blockdevice size update v2
    * block: blk_cbt_update_size add block_d

    110.1 since 042stab109.12:
    * block: blk_cbt_update_size add block_dev sanity check (PSBM-34151,
      #3240, #3257)
    * rebase of 042stab109.12 to RHEL6.6 2.6.32-504.23.4.el6 kernel
    * siee also RHSA-2015-1081   +++++++++++++++++++++++++++

    110.2 since 042stab110.1:
    * Double rtnl_unlock in netlink_dump() (#3270) (from 108.6)
    * diff-rst-iterative-adjust-percent-free-swap-pages (#3231)
    * cbt: ignore device shrink
    * ipv4: take rtnl_lock and mark mrt table as freed on namespace
    * extra rtnl_unlock in sky2_resume()
    * missing rtnl_unlock in i40evf_resume()
    * missing rtnl_unlock in igb_sriov_reinit()
    * rtnl_lock called twice in i40e_pci_error_resume()
    * rtnl_lock in ip6mr_net_exit()

    111.1 since 042stab110.2:
    * fix name_to_handle_at() for empty pathname (#3280)
    * AT_EMPTY_PATH support for fstatat() with empty relative (#3280)

    111.4 since 042stab111.1:
    * Fix for CVE-2015-5366 and CVE-2015-5364 extracted from RHEL6 kernel
    2.6.32-573.3.1.el6 backported version of mainline commit beb39db59d1
    ++++++++++++++++++
    "udp: fix behavior of wrong checksums".
    * See also: CVE-2015-5366 CVE-2015-5364

    111.5 since 042stab111.4:
    * tty: Fix recursive deadlock in tty_perform_flush() (PSBM-37612)
    * netfilter/ip6table_raw: Add permission check in ip6table_raw pernet (PSBM-35380)
    * ploop: dio_fastmap() must refresh bvec_merge_data
    * ploop: fix race on prealloc request (PSBM-35095)

    111.11 since 042stab111.5:
    * cpt: do not dump packet sock vma to pram (PSBM-39014)
    * sched: Fix schedule_tail() to disable preemption (PSBM-35082)
    * device_cgroup: fake allowing all devices for docker inside VZCT (PSBM-38691)
    * cbt: fix bytes to block conversion bug (PSBM-39050)
    * s/PTR_ERR/PTR_RET, otherwise we always go to the err path (PSBM-35380)
    * dm-thin: do not optimize requeued bios (PSBM-39205)
    * fixed host hang where ploop stuck inside ploop_make_request()
      (PSBM-29066, OVZ-6293, OVZ-6392)
    * cbt: add missed mutex_unlock (PSBM-39257)

    111.12 since 042stab111.11:
    * In certain circumstances, container owner could escape from
    bindmount. (PSBM-34244)

+++ ipc/shm.c
@@ -416,7 +416,7 @@ static int newseg(struct ipc_namespace *ns, struct ipc_params *params)
 	int shmid = params->id;
 	int error;
 	struct shmid_kernel *shp;
-	int numpages = (size + PAGE_SIZE -1) >> PAGE_SHIFT;
+	size_t numpages = (size + PAGE_SIZE - 1) >> PAGE_SHIFT;
 	struct file * file;
 	char name[64];
 	int id;
